Add skippable option.
Allow items to be marked as not freezable (a.k.a skippable during freezing).
Setting the `skippable` option would make the `freeze()` method
to check a boolean field for each model.
Set `skippable` to TRUE for the default field `is_not_freezable` or
set it to the name of the boolean field you want to check.
Freezable behavior would create the field as a boolean and pass `skippable_field_options`.